qcacmn: Fix logical_link_index for STA vaps
There was a case in which the logical_link_index for link vdev and
partner vdev was matching which was causing an issue.

The reason for duplicate logical link was because the mlo_link_connect
was not sent in same order in which partner link vdevs are added.

Send mlo_send_link_connect in order in which partner links are added.
This will ensure generation of logical link indices for link vdev
and partner vdevs correctly.

Change-Id: I4053fad7a7b615bca9b6386bf142ea785bca5ccf
CRs-Fixed: 3427985