qcacld-3.0: Fix incorrect wma_er() function call
A typo was introduced by the following patch:
  qcacld-3.0: Replace WMA_LOGE() with wma_err() in wma_main.c
  Change-Id If55d52640dbe921377f1ceffed609e096f449563

One instance of WMA_LOGE(), in wma_motion_det_host_event_handler(),
was changed to wma_er() instead of wma_err(). This has gone undetected
since it is feature-specific code that is not enabled, and hence has
not resulted in a compilation failure. But this was detected manually
during dead code removal, so fix the typo.

Change-Id: I07b6b5792c1419a83ee1a735943b8375d22e71d8
CRs-Fixed: 3432958