qcacld-3.0: Change to set proper tx power for Auto case
Tx power not set in case when user sets tx power as auto.
When user set tx power as auto now, host will configure
the tx power as max regulatory power allowed for
that channel. It signifies that it is the upper
limit for tx power used while transmission.

Change-Id: I98e53e417fa72c58ef8b22017a9a622ba8ccb257
CRs-Fixed: 3478243