qcacld-3.0: Send NAN PASN peer add and del req to target
Currently, host driver sends PASN peer create and delete request to
firmware for RTT peer only.
With new PASN implementation for NAN,  driver need to create and
delete the PASN peer and send the same request to the target.

So, to enable this, add support to send the NAN PASN peer create and
delete request.

Change-Id: I7bfd325a3b8de91ad3ee598faafb515dc83fb457
CRs-Fixed: 3774307