qcacld-3.0: Handle NAN PASN peer del response
When userspace send del key command on NAN interface to the driver,
driver will send peer delete command to the target. The target will
respond to the driver with peer delete response and driver will
remove the peer.
Currently, driver handles the PASN peer delete response for RTT
only.

So, to add support for the NAN PASN peer in host driver, handle
the peer delete response for NAN also.

Change-Id: Iebe18c0f73608856e7dbca2f6163c807c3daf86f
CRs-Fixed: 3774771