qcacld-3.0: Pack Supported channel IE based on range
In current design, each supported channel is packed
individually with range count as 1. This bloats the
size of the supported channel IE.

Instead as an optimization, break the supported channel
list into contiguous range of channels and encode only
the start frequency of the range and count of channels
in that range.

Change-Id: Ic5ba530ee6a8422ffc0c67acb621e76826775fa7
CRs-Fixed: 3467712