qcacld-3.0: Reject SAP start on 6 GHz channel if band bitmask not set
Currently host doesn't check if 6 GHz band bitmask is set
when receives SAP start on a 6 GHz channel. This can cause
SAP start on a 6 GHz channel when set_fcc_channel is set.

To avoid this issue, reject SAP start on a 6 GHz channel
if 6 GHz band bitmask is not set.

Change-Id: I0dd2b1fac7c4293498f7747665302bd72b7aaa23
CRs-Fixed: 3481002