qcacld-3.0: Remove duplicate DFS_INC inclusion
The compiler switches defined by DFS_INC are being included in the
composite INCS twice. This results in the switches being unnecessarily
duplicated in the compiler command line. While this normally doesn't
affect compilation, it can result in a "command too long" error that
has been observed in some build environments, so remove the
duplication.

Change-Id: Ibe09e7c3d4ba1a9b5b50305c73b79ce1c9a7d57a
CRs-Fixed: 3487038