qcacld-3.0: Sort Kconfig
Since the qcacld-3.0 driver is not part of the upstream kernel, the
Kconfig has not been used for driver configuration. As a result, it is
missing hundreds of configuration options that are now supported by
the driver.

The upcoming Bazel DDK build infrastructure *will* use the Kconfig
file. In preparation for populating the Kconfig, sort the existing
options in alphabetic order to facilitate the addition of the missing
options in a subsequent patch.

Change-Id: I7a714d373ae601b8c6642c224aca6b64da618ecd
CRs-Fixed: 3487378