qcacld-3.0: Reject connection if duplicate peer on same VDEV
The logic to allow peer create on same ML dev context is useful to allow
connection to ML-AP with same link address and MLD address, but it will
allow duplicate peer for ML-SAP.

If a legacy client is already connected to ML-SAP and new ML client
initiates connection with MLD address same as legacy client MAC address,
the rejection logic during peer create fails as both legacy and new ML
client belong to same ML dev context.

During peer create if a duplicate entry found in ML dev context, only
allow if the duplicate entry is on different VDEV, if it is on same VDEV
reject the peer create.

Change-Id: Ic82f80d814e21ea0b69ffd960e7aaef4ebc80ebf
CRs-Fixed: 3496058