qca-wifi-oss: Handle OCAC completion status in Agile SM
If configuring the RCAC channel fails or if the Agile RCAC is
aborted, OCAC completion event 'WMI_VDEV_ADFS_OCAC_COMPLETE_EVENTID'
is received from FW with status 'OCAC_CANCEL'. Currently, OCAC
completion is not handled based on its sub status ( OCAC_SUCCESS or
OCAC_CANCEL ).

Hence, if the OCAC completion is received out of sync with the current
DFS agile operation, the channel with non-success completion status
gets marked as agile RCAC done.

For example, Consider the case where agile RCAC is currently running
on a channel. Consider the user configures a new agile RCAC channel
or reconfigures the same channel. The following sequence of events can
cause a race condition.
-> The current host agile RCAC timer is stopped and RCAC abort command
   is sent to FW.
-> Agile RCAC is restarted with the newly configured channel.
-> But, by this time, the RCAC completion event is received from FW as
   a response to the abort command sent from host. This causes OCAC
   completion event to be posted to the SM, considering the OCAC
   completion is received for the current agile channel.
-> Since the agile SM does not check the status of the OCAC completion
   event, the channel gets incorrectly marked as RCAC done.

To solve this issue, move the OCAC completion sanity check into
the Agile SM, which is within a lock to make the processing of an
event atomic.

Also, decide the next state of the agile state machine based on the
status of the OCAC completion event.

Mark the channel as OCAC done only if the OCAC completion status is
success. If the status is 'OCAC_CANCEL', abort the agile RCAC timer
and restart the Agile RCAC.

Since the agile SM now takes into account the status of the OCAC
completion event, make changes across the agile SM to keep the agile
OCAC completion status consistent.
-> Initialise the OCAC status to 'OCAC_RESET'.
-> Mark the status as 'OCAC_SUCCESS' before posting the OCAC done
   event to the SM.
-> Reset the OCAC status after processing the OCAC completion event.

This change also replaces the OCAC status type from boolean to
the enumerated type to support the changes made in
"Id95014f4f81b4a2137092a5dd107c570d70058b6".

Change-Id: I1c694e22f80aa519f03b8c9cddd8e1c48e7ecbde