qca-wifi-oss: Configure state machine based on radar
When puncture SM is stopped, set dfs_is_unpunctured
variable to true. So that the state of SM is remembered.
This prevents from sending radar/NOL expiry/CAC expiry
events to an SM which is already punctured.

When radar is seen in a DFS channel, initialize the SM
object indexing to the punctured channel. The maximum
SM available is two. Currently iterating the SM objects
based on number of punctured channels. This cause issue
when two radars are seen at two different interval of
time, because in this case the number of punctured channel
at that one instance of radar is one and iterate through
one SM object only. Therefore the second punctured channel
is not allocated with SM object. Therefore always iterate
through all the SM objects and allocate them.

If the punctured channel list is needed, then generate
them based on the radar found channel instead of dfs
curchan. Some cases dfs_curchan can be different from
radar found channel. For example, if there is a channel
change triggered by user then dfs_curchan is modified
and it has no information about the punctured channels.

When radar is found, if the generated bitmap is zero,
dont initialize or start the puncture SM.

No need to allocate the punctured channel list with
the size of MAX_20MHZ_SUBCHANS, because the current
maximum punctured channels in a sub-band is two.
Therefore allocate two as the size of punctured channel
list.

During handling of the radar puncturing, always add
the existing radar puncture pattern to the current
new puncture pattern. Previously we allocate only
the current new puncture pattern and it forgets
the existing puncture pattern.

The DFS puncture SM is allocated based on number of
punctured channels. This cause issue when two radars
are injected at two different intervals. Instead,
allocate SM based on number of radars.

When channel is removed from NOL, check if the SM
object is allocated with low and high frequency value,
if its not allocated no need to send the NOL expiry
event.

In API is_found_in_cur_chan_punc_list, previously only
the frequency is sent to check if the freq is present
in the current channel list, this fails if there is
one radar and two punctured channel case. Send the
punctured list and check if it is present in current
channel list.

Separate user puncture pattern and radar puncture
pattern. After unpuncturing the radar pattern, add
user patter and unpunctured pattern as new pattern.
The new pattern is sent through CSA.

The internal_radar_pattern is not cleared after the
VAP moves to another channel. The internal radar
pattern is not needed as the puncture SM object
stores the information of radar punctured channels.

CRs-Fixed: 3514049
Change-Id: Ifedcd926dcd0f787ef2b0b02690f8d8ed6e2b34d